Split widget: collections header row
